# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Receives Shopify order webhooks and relays them to Klaviyo as Placed Order events, with retries.",
        "title": "Shopify Klaviyo Relay",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/api/events": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "List relay events",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "default": 1,
                        "description": "Page number; values below 1 are clamped to 1",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "maximum": 100,
                        "minimum": 1,
                        "type": "integer",
                        "default": 50,
                        "description": "Page size; clamped to [1,100]",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.eventsPage"
                        }
                    },
                    "500": {
                        "description": "internal error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/webhook/shopify/orders": {
            "post": {
                "security": [
                    {
                        "ShopifyHmac": []
                    }
                ],
                "description": "Stores the order for asynchronous relay to Klaviyo. The raw request body is capped at 1 MiB and must carry a valid HMAC signature.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Receive a Shopify order webhook",
                "parameters": [
                    {
                        "description": "Shopify order payload",
                        "name": "order",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.ShopifyOrder"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.webhookResponse"
                        }
                    },
                    "400": {
                        "description": "invalid payload",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "internal error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.LineItem": {
            "type": "object",
            "properties": {
                "price": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "api.ShopifyOrder": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "currency": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "line_items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.LineItem"
                    }
                },
                "name": {
                    "type": "string"
                },
                "total_price": {
                    "type": "string"
                }
            }
        },
        "api.eventResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "last_error": {
                    "type": "string"
                },
                "order_id": {
                    "type": "integer"
                },
                "retry_count": {
                    "type": "integer"
                },
                "status": {
                    "type": "string",
                    "enum": [
                        "received",
                        "retrying",
                        "succeeded",
                        "failed",
                        "expired"
                    ]
                }
            }
        },
        "api.eventsPage": {
            "type": "object",
            "properties": {
                "events": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.eventResponse"
                    }
                },
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "api.webhookResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "format": "uuid"
                }
            }
        }
    },
    "securityDefinitions": {
        "ShopifyHmac": {
            "description": "Base64 HMAC-SHA256 signature of the raw request body, computed with the shared webhook secret. Not a static key — Swagger UI cannot generate it; use scripts/send_webhook.sh to send a signed request.",
            "type": "apiKey",
            "name": "X-Shopify-Hmac-SHA256",
            "in": "header"
        }
    }
}